rename FormSkeleton file to Skeleton
