+++ catalog/app/containers/Bucket/PackageDialog/Skeleton.js
